requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
